DESeq2 differential expression analysis

Class containing all elements generated during differential expression analysis with DESeq2. This class is essentially a list with validity checks to ensure DESeqTransform and DESeqResults correspond to the DESeqDataSet.

Value

DESeqAnalysis.

Slots

data

DESeqDataSet.

transform

DESeqTransform.

results

list. One or more unshrunken DESeqResults.

lfcShrink

list. Optional. One or more shrunken DESeqResults. If set, must correspond to those defined in results. Otherwise, can set as empty list (list()).

Note

Updated 2019-10-15.

DESeqDataSet

We recommend generating the DESeqDataSet by coercion from bcbioRNASeq object using as(dds, "bcbioRNASeq"). Don't use the DESeq2::DESeqDataSet() or DESeq2::DESeqDataSetFromMatrix() constructors to generate the DESeqDataSet object.

DESeqTransform

Object containing variance-stabilized counts. We recommend slotting the return from either DESeq2::varianceStabilizingTransformation() or DESeq2::rlog().

DESeqResults

Don't modify any of the DESeqResults objects manually. This includes rearranging the rows or dropping genes without adjusted P values. We'll take care of this automatically in supported methods.

See also