DESeq2 Differential Expression Results Tables

DESeqResults object with DataFrame subsets and file path information.

DESeqResultsTables(object, ...)

# S4 method for DESeqAnalysis
DESeqResultsTables(object, results = 1L,
  lfcShrink = TRUE)

Arguments

object
Object.

...
Additional arguments.

results
scalar. Position or name of DESeqResults.

lfcShrink
boolean. Use shrunken log2 fold change (LFC) values.

Value

DESeqResultsTables, containing the original, unmodified DESeqResults DataFrame along with the corresponding differentially expressed genes and gene-level metadata (rowRanges).

Obtaining results from DESeq2

It is recommended to specify the contrast argument as character:

Design matrix factor of interest.

Numerator for LFC (expt).

Denominator for LFC (control).

For example, to get the relative change in gene expression between mutant and wildtype samples, here's how to set the contrast vector:

c(
    factor = "genotype",
    numerator = "mutant",
    denominator = "wildtype"
)

Log fold change threshold cutoffs

Refer to DESeq2::results() for additional information about using lfcThreshold and altHypothesis to set an alternative hypothesis based on expected fold changes. In addition, the "Hypothesis tests with thresholds on effect size" section in the DESeq2 paper provides additional explanation.

Don't apply post hoc LFC threshold filtering to obtain results tables, as this will destroy the meaning of the adjusted P values computed. If you are expecting a biological effect size past a particular threshold, set lfcThreshold, but be conservative.

This thread on the Bioconductor forums explains how DESeq2::results() should be called with regard to LFC cutoffs in nice detail.

See also

Examples

data(deseq)

## DESeqAnalysis ====
x <- DESeqResultsTables(deseq)

#> DESeqResults: condition B vs A (shrunken LFC)

print(x)

#> DESeqResultsTables 1.22.1
#> contrast: condition B vs A
#> alpha: 0.1
#> lfcThreshold: 0
#> out of 499 with nonzero total read count
#> adjusted p-value < 0.1
#> LFC > 0 (up)       : 84, 17%
#> LFC < 0 (down)     : 87, 17%
#> outliers [1]       : 2, 0.4%
#> low counts [2]     : 0, 0%
#> (mean count < 0)