Alpha level cutoff summary statistics

Quickly generate a summary table of various alpha level cutoffs.

alphaSummary(object, ...)

# S4 method for DESeqDataSet
alphaSummary(
  object,
  alpha = c(0.1, 0.05, 0.01, 0.001, 1e-06),
  contrast = NULL,
  name = NULL
)

# S4 method for DESeqAnalysis
alphaSummary(object, ...)

Arguments

object
Object.

alpha
numeric. Multiple alpha cutoffs.

contrast
character. A character vector with exactly 3 elements:

Name of factor in the design formula.

Name of numerator level for the fold change.

Name of denominator level for the fold change.

See DESeq2::results() for details.

name
character(1). Name of the individual effect (coefficient) for building a results table. Use this argument rather than contrast for continuous variables.

...
Additional arguments.

Value

integer matrix.

Details

Use either contrast or name to specify the desired contrast.

Note

Updated 2019-08-20.

See also

Author

Michael Steinbaugh, Lorena Patano

Examples

data(deseq)

## DESeqAnalysis ====
alphaSummary(deseq, contrast = c("condition", "B", "A"))

#> ℹ condition B A

#>                0.1 0.05 0.01 0.001 1e-06
#> LFC > 0 (up)    48   36   24    15     2
#> LFC < 0 (down)  39   22   14     8     1
#> outliers [1]     0    0    0     0     0
#> low counts [2] 116  116   58   271     0

alphaSummary(deseq, name = "condition_B_vs_A")

#> ℹ condition_B_vs_A

#>                0.1 0.05 0.01 0.001 1e-06
#> LFC > 0 (up)    48   36   24    15     2
#> LFC < 0 (down)  39   22   14     8     1
#> outliers [1]     0    0    0     0     0
#> low counts [2] 116  116   58   271     0