Differentially Expressed Genes

Differentially Expressed Genes

# S4 method for DESeqResults
deg(object, alpha = NULL, lfcThreshold = NULL,
  direction = c("both", "up", "down"))

# S4 method for DESeqAnalysis
deg(object, results = 1L,
  direction = c("both", "up", "down"))

Arguments

object
Object.

alpha
numeric(1). Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

lfcThreshold
numeric(1). Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

direction
character(1). Plot "both", "up", or "down" directions.

results
character(1) or integer(1). Name or positive of DESeqResults.

Value

character. Gene identifiers.

Examples

data(deseq)

## DESeqAnalysis ====
deg(deseq)

#> 171 differentially expressed genes detected.

#>   [1] "gene427" "gene419" "gene095" "gene334" "gene323" "gene494" "gene352"
#>   [8] "gene391" "gene145" "gene078" "gene403" "gene497" "gene248" "gene004"
#>  [15] "gene285" "gene489" "gene063" "gene023" "gene148" "gene448" "gene215"
#>  [22] "gene473" "gene412" "gene060" "gene105" "gene277" "gene298" "gene317"
#>  [29] "gene291" "gene363" "gene181" "gene097" "gene166" "gene217" "gene441"
#>  [36] "gene449" "gene456" "gene364" "gene001" "gene375" "gene085" "gene321"
#>  [43] "gene235" "gene198" "gene117" "gene405" "gene378" "gene366" "gene422"
#>  [50] "gene377" "gene395" "gene439" "gene057" "gene150" "gene491" "gene206"
#>  [57] "gene460" "gene320" "gene336" "gene116" "gene444" "gene374" "gene091"
#>  [64] "gene003" "gene360" "gene270" "gene495" "gene384" "gene376" "gene388"
#>  [71] "gene394" "gene299" "gene385" "gene250" "gene333" "gene338" "gene305"
#>  [78] "gene021" "gene365" "gene209" "gene272" "gene269" "gene223" "gene498"
#>  [85] "gene133" "gene469" "gene236" "gene415" "gene470" "gene231" "gene082"
#>  [92] "gene160" "gene090" "gene355" "gene300" "gene058" "gene396" "gene039"
#>  [99] "gene142" "gene176" "gene478" "gene122" "gene417" "gene500" "gene044"
#> [106] "gene208" "gene274" "gene140" "gene214" "gene156" "gene327" "gene149"
#> [113] "gene273" "gene067" "gene358" "gene483" "gene075" "gene304" "gene411"
#> [120] "gene045" "gene315" "gene292" "gene123" "gene111" "gene141" "gene408"
#> [127] "gene342" "gene340" "gene071" "gene399" "gene143" "gene167" "gene080"
#> [134] "gene496" "gene454" "gene012" "gene468" "gene204" "gene341" "gene035"
#> [141] "gene129" "gene393" "gene447" "gene014" "gene401" "gene252" "gene022"
#> [148] "gene261" "gene283" "gene329" "gene138" "gene186" "gene265" "gene354"
#> [155] "gene147" "gene325" "gene087" "gene175" "gene418" "gene369" "gene177"
#> [162] "gene301" "gene066" "gene295" "gene486" "gene114" "gene475" "gene069"
#> [169] "gene330" "gene350" "gene428"