Intersection of differentially expressed genes

Intersection of differentially expressed genes

degIntersection(object, ...)

# S4 method for DESeqAnalysis
degIntersection(
  object,
  i = NULL,
  direction = c("up", "down"),
  return = c("matrix", "count", "ratio")
)

Arguments

object
Object.

i
character, numeric, or NULL. Names or range of results. If set NULL, include all results.

direction
character(1). Include "up" or "down" directions. Must be directional, and intentionally does not support "both".

return
character(1). Return intersection matrix, count per contrast, or ratio.

...
Passthrough arguments to deg().

Value

integer. Integer denoting the number of intersections, sorted from highest to lowest. Gene identifiers are defined as the names.

Note

Updated 2020-08-18.

Examples

data(deseq)

## DESeqAnalysis ====
mat <- degIntersection(deseq, return = "matrix")

#> → Returning intersection matrix of 24 up-regulated DEGs.

head(mat)

#>         condition_B_vs_A treatment_D_vs_C
#> gene027             TRUE            FALSE
#> gene028             TRUE            FALSE
#> gene030             TRUE            FALSE
#> gene057             TRUE            FALSE
#> gene109             TRUE            FALSE
#> gene112             TRUE            FALSE

count <- degIntersection(deseq, return = "count")

#> → Returning intersection count of 24 up-regulated DEGs.

head(count)

#> gene027 gene028 gene030 gene057 gene109 gene112 
#>       1       1       1       1       1       1 

ratio <- degIntersection(deseq, return = "ratio")

#> → Returning intersection ratio of 24 up-regulated DEGs.

head(ratio)

#> gene027 gene028 gene030 gene057 gene109 gene112 
#>     0.5     0.5     0.5     0.5     0.5     0.5