Export

Export data out of R and write to disk.

export(object, ...)

# S4 method for DESeqDataSet
export(object, name = NULL, dir = ".", compress = FALSE)

# S4 method for DESeqAnalysis
export(object, name = NULL, dir = ".", compress = FALSE)

Arguments

object
Object.

name
character(1). Name to use on disk. If NULL, will use the name of the object instead.

dir
character(1). Directory path.

compress
logical(1). Apply gzip compression to all files.

...
Additional arguments.

Value

Invisibily return file paths, either as list or character.

Note

Size-factor normalized coutns and FPKM values are calculated on the fly and exported automatically.

Updated 2020-08-04.

Examples

data(deseq)

## DESeqAnalysis ====
export(deseq, dir = "example")

#> → Exporting `DESeqDataSet` to data.

#> → Exporting `data` to /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data.

#> → Exporting assays `counts, normalized, fpkm` to /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data.

#> → Exporting counts.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data/assays using vroom::`vroom_write()`.

#> → Exporting normalized.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data/assays using vroom::`vroom_write()`.

#> → Exporting fpkm.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data/assays using vroom::`vroom_write()`.

#> → Exporting colData.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data using vroom::`vroom_write()`.

#> Dropping non-atomic columns: entrezID.

#> → Exporting rowData.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/data using vroom::`vroom_write()`.

#> → Exporting `DESeqTransform` to transform.

#> → Exporting `transform` to /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/transform.

#> → Exporting assays `assay` to /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/transform.

#> → Exporting assay.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/transform/assays using vroom::`vroom_write()`.

#> → Exporting colData.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/transform using vroom::`vroom_write()`.

#> Dropping non-atomic columns: entrezID.

#> → Exporting rowData.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/transform using vroom::`vroom_write()`.

#> → Exporting `DESeqResults` tables to resultsTables.

#> ℹ condition_B_vs_A (shrunken LFC)

#> ! Returning with the sample names unmodified.

#> ℹ 38 differentially expressed genes (alpha < 0.01)

#> ℹ 24 upregulated genes (alpha < 0.01)

#> ℹ 14 downregulated genes (alpha < 0.01)

#> → Exporting all.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsTables/condition_B_vs_A using vroom::`vroom_write()`.

#> → Exporting up.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsTables/condition_B_vs_A using vroom::`vroom_write()`.

#> → Exporting down.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsTables/condition_B_vs_A using vroom::`vroom_write()`.

#> → Exporting both.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsTables/condition_B_vs_A using vroom::`vroom_write()`.

#> ℹ treatment_D_vs_C (shrunken LFC)

#> ! Returning with the sample names unmodified.

#> ℹ 0 differentially expressed genes (alpha < 0.01)

#> → Exporting all.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsTables/treatment_D_vs_C using vroom::`vroom_write()`.

#> → Exporting `DESeqResults` matrices to resultsMatrices.

#> → Exporting log2FoldChange.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsMatrices using vroom::`vroom_write()`.

#> → Exporting stat.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsMatrices using vroom::`vroom_write()`.

#> → Exporting padj.csv at /Users/mike/git/monorepo/r-packages/DESeqAnalysis/docs/reference/example/deseq/resultsMatrices using vroom::`vroom_write()`.

## Clean up.
unlink("example", recursive = TRUE)