Differentially expressed gene heatmap

This function is an extension of plotHeatmap() that is optimized for automatic handling differentially expressed genes, rather than requiring manual input of a gene vector. All of the optional parameters for plotHeatmap() are also supported by this function.

plotDEGHeatmap(object, ...)

# S4 method for DESeqResults
plotDEGHeatmap(
  object,
  DESeqTransform,
  alphaThreshold = NULL,
  lfcThreshold = NULL,
  baseMeanThreshold = NULL,
  direction = c("both", "up", "down"),
  title = TRUE,
  subtitle = TRUE,
  ...
)

# S4 method for DESeqAnalysis
plotDEGHeatmap(object, i, contrastSamples = FALSE, ...)

Arguments

object
Object.

DESeqTransform
DESeqTransform.

alphaThreshold
numeric(1) or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

lfcThreshold
numeric(1) or NULL. Log (base 2) fold change ratio cutoff threshold. If left NULL, will use the cutoff defined in the object.

baseMeanThreshold
numeric(1) or NULL. Base mean (i.e. average expression across all samples) threshold. If left NULL, will use the cutoff defined in the object. Applies in general to DESeq2 RNA-seq differential expression output.

direction
character(1). Include "both", "up", or "down" directions.

title
logical(1), character(1). Include contrast name as title? Can manually define as character.

subtitle
logical(1). Include subtitle containing DEG information?

...
Additional arguments.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

contrastSamples
logical(1). Experimental. Only include the samples used to define the contrast passed to DESeq2::results(). This setting will break for complex DESeq2 contrasts (e.g. interaction effect).

Value

Plot.

Details

To adjust the annotation columns, modify the colData() of the counts argument, which must contain a SummarizedExperiment.

Functions

plotDEGHeatmap,DESeqResults-method: Passes to plotHeatmap() SummarizedExperiment method defined in AcidPlots.

plotDEGHeatmap,DESeqAnalysis-method: Passes to DESeqResults method.

Note

Updated 2020-08-05.

Examples

data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq, i = 1L)

#> Error in plotDEGHeatmap(deseq, i = 1L): could not find function "plotDEGHeatmap"