Relative difference of DESeq results

Relative difference of DESeq results

resultsDiff(x, y, ...)

# S4 method for DESeqResults,DESeqResults
resultsDiff(
  x,
  y,
  col = c("log2FoldChange", "stat"),
  deg = c("no", "intersect", "union"),
  alpha = NULL,
  lfcThreshold = NULL,
  direction = c("both", "up", "down")
)

# S4 method for DESeqAnalysis,missingOrNULL
resultsDiff(x, y = NULL, i, j, lfcShrink = TRUE, ...)

Arguments

x
Object.

y
Object.

...
Additional arguments.

col
character(1). Column to use for difference calculation (subtraction). Note that x/i represents A and y/j B in A - B calculation.

deg
character(1).

"no": Return all genes, corresponding to original row names order.

"intersect": Return only genes called as DE in both contrasts.

"union": Return genes called as DE in either contrast.

alpha
numeric(1) or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

lfcThreshold
numeric(1) or NULL. Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

direction
character(1). Plot "both", "up", or "down" directions.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

j
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

lfcShrink
logical(1). Use shrunken log2 fold change (LFC) values.

Value

Named numeric. Names correspond to the genes (row names).

Functions

resultsDiff,DESeqAnalysis,missingOrNULL-method: Passes arguments to DESeqResults method.

Note

Updated 2019-12-17.

Examples

data(deseq)

## DESeqResults ====
x <- results(deseq, i = 1L)

#> condition_B_vs_A (unshrunken LFC)

y <- results(deseq, i = 2L)

#> treatment_D_vs_C (unshrunken LFC)

resultsDiff(x = x, y = y)

#> Calculating relative difference of 'log2FoldChange'.

#>      gene001      gene002      gene003      gene004      gene005      gene006 
#> -2.147588403  0.328056375 -2.885499319 -1.186228332 -0.055237120  0.284704058 
#>      gene007      gene008      gene009      gene010      gene011      gene012 
#>  0.975158574  1.894691376  1.478910199  1.081824304 -0.594282979 -0.030791322 
#>      gene013      gene014      gene015      gene016      gene017      gene018 
#> -1.196213248 -0.509927744  1.585193616  1.701270163 -2.019533288  0.605710135 
#>      gene019      gene020      gene021      gene022      gene023      gene024 
#>  0.726981738  2.352143680 -0.624582174 -5.157547574  0.141328903 -1.580432323 
#>      gene025      gene026      gene027      gene028      gene029      gene030 
#> -1.906237491 -0.992154407  0.857422887  0.852382028 -0.653657999  0.437621574 
#>      gene031      gene032      gene033      gene034      gene035      gene036 
#>  1.031045131  2.339228882  3.419751285 -2.203021394  2.271918839 -1.460828816 
#>      gene037      gene038      gene039      gene040      gene041      gene042 
#>  3.699553729  0.009122654 -0.616538268 -0.539631105  0.299426511  0.761436364 
#>      gene043      gene044      gene045      gene046      gene047      gene048 
#> -2.620877398 -0.513696515  0.520060455  0.057107594 -0.040852876  0.822510661 
#>      gene049      gene050      gene051      gene052      gene053      gene054 
#>  0.109204244  2.567873986 -1.368504591 -1.362212679  1.065700067 -3.071919263 
#>      gene055      gene056      gene057      gene058      gene059      gene060 
#> -1.267339869  0.048419244 -2.819258677  0.868058352  1.478948942  0.605696589 
#>      gene061      gene062      gene063      gene064      gene065      gene066 
#>  1.141631736  1.462139650  0.330063441 -3.750614330 -2.895750828  0.865550749 
#>      gene067      gene068      gene069      gene070      gene071      gene072 
#>  1.862431362  0.019571893 -0.427019020  0.036108209 -0.118280933  0.564974888 
#>      gene073      gene074      gene075      gene076      gene077      gene078 
#>  0.987738668 -0.677228279 -1.419074333  1.017332887 -3.620635200 -0.118166369 
#>      gene079      gene080      gene081      gene082      gene083      gene084 
#> -0.598495551  0.092121063 -0.283387649  0.823649432  1.883621117 -1.738653142 
#>      gene085      gene086      gene087      gene088      gene089      gene090 
#>  1.132454273  1.376748350  0.258954239  1.171857171 -2.660174307  1.463655326 
#>      gene091      gene092      gene093      gene094      gene095      gene096 
#>  0.408729221  0.463170233  0.150911975  0.741050459  1.453220667  1.518759787 
#>      gene097      gene098      gene099      gene100      gene101      gene102 
#> -1.711034883  1.631614307 -1.294178711 -0.985908917 -0.123374524  0.840380119 
#>      gene103      gene104      gene105      gene106      gene107      gene108 
#> -1.541741829  0.252511264  3.841535098 -0.519606380 -0.573532630  0.679079075 
#>      gene109      gene110      gene111      gene112      gene113      gene114 
#>  1.050097785 -2.120875313  0.644451395 -1.440461633 -0.796657044  3.105526949 
#>      gene115      gene116      gene117      gene118      gene119      gene120 
#>  0.680080291  2.187961782  2.139824718  0.598129074  0.339782172 -0.359004776 
#>      gene121      gene122      gene123      gene124      gene125      gene126 
#>  0.703093731 -1.564042518 -0.368762203 -1.294130075  2.687096584 -0.382505080 
#>      gene127      gene128      gene129      gene130      gene131      gene132 
#> -0.760381230  0.608597553 -1.983192855 -0.034859372 -0.427402055 -0.121283409 
#>      gene133      gene134      gene135      gene136      gene137      gene138 
#> -4.288467629 -1.236696206 -1.821689049  1.423649187 -2.514783813  1.233980540 
#>      gene139      gene140      gene141      gene142      gene143      gene144 
#>  1.123415918 -3.398808012  1.046129115  1.196965293 -2.375153597 -0.872743674 
#>      gene145      gene146      gene147      gene148      gene149      gene150 
#>  1.516297223 -1.450270007 -1.907784823 -0.348447971 -0.246407631 -0.597579223 
#>      gene151      gene152      gene153      gene154      gene155      gene156 
#> -2.019646858  1.630811240  0.106632711  0.137496668 -0.669158046 -2.635209134 
#>      gene157      gene158      gene159      gene160      gene161      gene162 
#> -0.732733240 -1.785126819 -0.629810251  1.557172308 -0.916910199  0.866425222 
#>      gene163      gene164      gene165      gene166      gene167      gene168 
#>  1.329861712 -1.559974950 -1.667361413 -0.099264417 -0.560389967  0.546755660 
#>      gene169      gene170      gene171      gene172      gene173      gene174 
#>  2.420575293  0.057805846  3.195814775  0.449325269 -2.590980758 -0.309492646 
#>      gene175      gene176      gene177      gene178      gene179      gene180 
#>  0.603211704 -4.251643972 -0.087373407 -3.185840024 -0.082091417  2.884955394 
#>      gene181      gene182      gene183      gene184      gene185      gene186 
#> -0.621266509  0.503965820 -0.057959437 -1.683519283 -3.024459282  1.262576951 
#>      gene187      gene188      gene189      gene190      gene191      gene192 
#> -0.110643283 -3.265696335 -3.036977612  2.663243961 -1.894963517 -1.316712360 
#>      gene193      gene194      gene195      gene196      gene197      gene198 
#>  0.891690515  0.482237146  1.284770239 -1.539051023 -0.089749853 -1.433304663 
#>      gene199      gene200      gene201      gene202      gene203      gene204 
#>  0.383449128  1.402197762 -2.430888582 -1.037485140 -2.004122443 -0.368342114 
#>      gene205      gene206      gene207      gene208      gene209      gene210 
#> -0.671055384  0.040542919  1.654621695  0.149919340 -1.182759783 -1.869141816 
#>      gene211      gene212      gene213      gene214      gene215      gene216 
#> -1.685704703 -0.413615284 -2.026785063 -1.356935861  0.938807921  1.594179465 
#>      gene217      gene218      gene219      gene220      gene221      gene222 
#>  0.915762063  0.101152787  1.173749300 -0.243499138  1.590168804 -0.839542274 
#>      gene223      gene224      gene225      gene226      gene227      gene228 
#> -0.121407233  0.132109492  0.843742338  1.397017609 -0.684607375 -0.046013305 
#>      gene229      gene230      gene231      gene232      gene233      gene234 
#>  0.805327645 -0.522947137  3.184903660 -0.117010230 -1.139274293  2.007906930 
#>      gene235      gene236      gene237      gene238      gene239      gene240 
#> -0.689788959 -0.091518573 -1.111297232 -0.927391436 -2.110498863 -1.262285320 
#>      gene241      gene242      gene243      gene244      gene245      gene246 
#> -0.560850164  2.008295674  1.385916480  2.321806416  0.997220544  1.385411964 
#>      gene247      gene248      gene249      gene250      gene251      gene252 
#>  0.937721820  1.632914243 -0.913141097 -0.665740748  0.002743292  2.216075591 
#>      gene253      gene254      gene255      gene256      gene257      gene258 
#>  0.020685376 -1.365213028 -0.037782990 -0.121254360  0.631643170  2.382660694 
#>      gene259      gene260      gene261      gene262      gene263      gene264 
#>  0.572880693 -1.491808792  0.840380119  1.471414720 -0.202013079 -0.317266837 
#>      gene265      gene266      gene267      gene268      gene269      gene270 
#>  0.474451585  1.935023729 -0.982979456  0.713971695  0.604289045 -3.209481362 
#>      gene271      gene272      gene273      gene274      gene275      gene276 
#> -1.353658050  1.457024147  1.876240680 -0.983159869  0.019244934  0.369646402 
#>      gene277      gene278      gene279      gene280      gene281      gene282 
#>  0.589314586 -0.093550923 -1.004202785 -2.010312935 -1.143940851 -2.989222730 
#>      gene283      gene284      gene285      gene286      gene287      gene288 
#>  0.096071344 -1.450617572 -0.202005717 -0.958604381  1.269709769 -0.549631840 
#>      gene289      gene290      gene291      gene292      gene293      gene294 
#>  0.056369689  1.456409243  0.626538122  0.642228344 -3.896661103 -0.972237013 
#>      gene295      gene296      gene297      gene298      gene299      gene300 
#>  0.511439148 -1.954962390  1.554236002  0.466368718 -1.055031821  1.472651637 
#>      gene301      gene302      gene303      gene304      gene305      gene306 
#>  0.832552904  0.450870279  0.540131048  1.144112512 -1.321228614  1.037776625 
#>      gene307      gene308      gene309      gene310      gene311      gene312 
#>  1.188717948  0.195794275 -0.448710726 -1.289407513 -0.085965849  0.665978974 
#>      gene313      gene314      gene315      gene316      gene317      gene318 
#>  1.316981893 -0.081023213  0.200682266 -1.116714285  0.367345431 -0.469182641 
#>      gene319      gene320      gene321      gene322      gene323      gene324 
#>  0.456981604  3.273081783 -0.907513423  0.879913836 -4.197039900  1.068542281 
#>      gene325      gene326      gene327      gene328      gene329      gene330 
#>  1.400881065  1.085535687 -1.512223150 -0.432912624 -0.162544710 -1.535237905 
#>      gene331      gene332      gene333      gene334      gene335      gene336 
#> -1.035979509  1.132731824 -0.732928392 -1.191445329 -1.751696269  0.692244923 
#>      gene337      gene338      gene339      gene340      gene341      gene342 
#>  0.975327462 -1.319707930 -1.282669064  1.982072471 -0.532328841  1.325239733 
#>      gene343      gene344      gene345      gene346      gene347      gene348 
#>  3.200967949  1.365725622  0.016552035 -0.545699193 -0.201795121  3.025740492 
#>      gene349      gene350      gene351      gene352      gene353      gene354 
#> -1.214202432 -0.077659367 -0.121407233 -0.542712120 -0.792213031  4.343289585 
#>      gene355      gene356      gene357      gene358      gene359      gene360 
#>  0.749152296 -1.122897986  0.398414611 -0.135975629  0.271787265 -1.933265576 
#>      gene361      gene362      gene363      gene364      gene365      gene366 
#> -1.790602600  0.312723068  1.443336163 -0.947009113 -2.409920764  0.061630109 
#>      gene367      gene368      gene369      gene370      gene371      gene372 
#>  0.659782147  2.022496528 -0.680757238  0.875821475  3.200304642  1.151148196 
#>      gene373      gene374      gene375      gene376      gene377      gene378 
#> -2.386988850  0.596047250 -2.334175074 -0.542221733  5.774702816 -0.643497822 
#>      gene379      gene380      gene381      gene382      gene383      gene384 
#> -0.460100445 -0.175260604 -0.542053378 -1.112427298 -0.626459916  1.069670715 
#>      gene385      gene386      gene387      gene388      gene389      gene390 
#> -1.391368944 -0.747351700  1.330375956  0.093188193  0.076179232 -0.652956106 
#>      gene391      gene392      gene393      gene394      gene395      gene396 
#>  0.176796021  0.797674955 -0.119072537  1.151429842 -1.757329693  1.015050247 
#>      gene397      gene398      gene399      gene400      gene401      gene402 
#> -0.372123066  0.050546635  0.483592502 -0.128635256 -0.086671521 -1.992990067 
#>      gene403      gene404      gene405      gene406      gene407      gene408 
#> -0.461951222 -2.853561979 -0.547322894 -2.281436108  1.362478072  0.862007471 
#>      gene409      gene410      gene411      gene412      gene413      gene414 
#>  1.709512473  0.913173807 -0.441148059 -1.145250097 -1.719478634 -1.075926359 
#>      gene415      gene416      gene417      gene418      gene419      gene420 
#> -0.394310294  0.730957909  2.337118598 -1.472899241  1.420126345 -0.266836406 
#>      gene421      gene422      gene423      gene424      gene425      gene426 
#>  0.093519143  1.889832940  0.053824716 -0.122793838  0.766781243 -0.758170163 
#>      gene427      gene428      gene429      gene430      gene431      gene432 
#>  2.975576766 -0.123215052 -0.489848460 -2.302931318  2.841093328  0.136776518 
#>      gene433      gene434      gene435      gene436      gene437      gene438 
#> -0.037620317 -1.604466487 -2.085255046  1.764053937 -0.217504773  0.579400979 
#>      gene439      gene440      gene441      gene442      gene443      gene444 
#> -0.121407768  3.305359023 -2.173502148 -2.383064355  1.950421838  1.097927986 
#>      gene445      gene446      gene447      gene448      gene449      gene450 
#> -0.121751678 -1.126813299 -3.207401199  2.689941690 -0.121407233 -0.088367745 
#>      gene451      gene452      gene453      gene454      gene455      gene456 
#>  0.342453843  0.400663859  2.675224178 -0.021833277 -0.103666646 -2.526117415 
#>      gene457      gene458      gene459      gene460      gene461      gene462 
#>  1.171744030 -2.031959295  3.194260714  2.656659824 -0.533961688  1.044238594 
#>      gene463      gene464      gene465      gene466      gene467      gene468 
#>  0.147094598 -1.735117574 -0.122081879 -2.066643577 -3.677612003 -0.783114666 
#>      gene469      gene470      gene471      gene472      gene473      gene474 
#>  0.873738374  1.995884264 -0.886839940 -1.751477892  0.064864337 -1.390259528 
#>      gene475      gene476      gene477      gene478      gene479      gene480 
#> -1.686683901 -1.011756492  0.452831436 -0.692489055 -2.984660917  2.995725203 
#>      gene481      gene482      gene483      gene484      gene485      gene486 
#> -1.186463700  0.286879566  2.678916898 -1.935116053  0.202527822  0.298326822 
#>      gene487      gene488      gene489      gene490      gene491      gene492 
#>  1.353302970  1.808471313 -5.609493103 -1.235731935  0.478558458  0.840380119 
#>      gene493      gene494      gene495      gene496      gene497      gene498 
#> -0.456855179  2.726519187  1.131029157 -1.130572966 -0.112598993 -0.172864867 
#>      gene499      gene500 
#> -2.511026968  1.385039542 

## DESeqAnalysis ====
object <- deseq
resultsDiff(object, i = 1L, j = 2L)

#> condition_B_vs_A (shrunken LFC)

#> treatment_D_vs_C (shrunken LFC)

#> Calculating relative difference of 'log2FoldChange'.

#>      gene001      gene002      gene003      gene004      gene005      gene006 
#> -0.748467715 -0.636028161 -1.816650847 -1.420572438 -0.346128020 -0.150649164 
#>      gene007      gene008      gene009      gene010      gene011      gene012 
#>  0.264703769 -0.062716893 -0.056258952  0.401130770 -0.287480220 -0.491241726 
#>      gene013      gene014      gene015      gene016      gene017      gene018 
#> -0.645882862 -0.098978023  0.446593493 -0.309626931 -0.161475887 -0.357071704 
#>      gene019      gene020      gene021      gene022      gene023      gene024 
#>  0.099220980  1.223190227 -0.197683137 -0.422902704  0.484964991  0.029443244 
#>      gene025      gene026      gene027      gene028      gene029      gene030 
#> -1.198051494 -0.650190524  0.467146803  0.789743801 -0.392778185  0.065856328 
#>      gene031      gene032      gene033      gene034      gene035      gene036 
#>  0.794757201  2.440426687  1.888657529 -1.334105387  2.429826271 -0.827565822 
#>      gene037      gene038      gene039      gene040      gene041      gene042 
#>  0.786340470  0.413840199 -0.442025123 -0.630530220  0.777976912  0.062834515 
#>      gene043      gene044      gene045      gene046      gene047      gene048 
#> -0.384736827 -0.320132241  0.643288512  0.032107519 -0.524598487  0.781660816 
#>      gene049      gene050      gene051      gene052      gene053      gene054 
#> -0.259987351  0.834710726 -0.924034627 -0.666085287  0.279918624 -0.478583008 
#>      gene055      gene056      gene057      gene058      gene059      gene060 
#> -0.932992109 -0.204881589 -0.243017027  0.165011187  1.395476872  0.727317648 
#>      gene061      gene062      gene063      gene064      gene065      gene066 
#>  0.962822117  1.549893384  0.154587699 -0.096444243 -2.192076140  0.444761012 
#>      gene067      gene068      gene069      gene070      gene071      gene072 
#>  1.440672131  0.590115236 -0.266923862  0.105193081 -0.160703433  0.401479128 
#>      gene073      gene074      gene075      gene076      gene077      gene078 
#>  0.667073564 -0.022416122 -0.281384172  0.479301355 -0.320757716  0.174558865 
#>      gene079      gene080      gene081      gene082      gene083      gene084 
#> -0.786589073  0.174239108  0.066368689  0.092511107  0.513042009 -0.956057687 
#>      gene085      gene086      gene087      gene088      gene089      gene090 
#>  2.053951884  0.709276234  0.522726593  0.028710757 -1.688882526  0.038618294 
#>      gene091      gene092      gene093      gene094      gene095      gene096 
#>  0.009807904  0.634725187  1.324269675 -0.019203265  0.309714756  1.376652973 
#>      gene097      gene098      gene099      gene100      gene101      gene102 
#> -0.795905458  0.242482086 -0.543773059 -0.900254552  0.078454754  0.067415096 
#>      gene103      gene104      gene105      gene106      gene107      gene108 
#>  0.318050948 -0.149029836  0.015574012 -0.104237911  0.143846570  1.181720061 
#>      gene109      gene110      gene111      gene112      gene113      gene114 
#>  0.487603325 -1.370562713  0.748442504 -0.612294732 -0.325049478  1.948588445 
#>      gene115      gene116      gene117      gene118      gene119      gene120 
#>  0.058276278  0.045972726  0.827345661  0.475259594 -0.041476825 -0.118761018 
#>      gene121      gene122      gene123      gene124      gene125      gene126 
#>  0.104078515 -0.342865135 -0.960445337 -1.299239703  0.249494399  0.051076814 
#>      gene127      gene128      gene129      gene130      gene131      gene132 
#> -0.410350431  0.364234634 -0.268906136 -0.334918167  0.319478147  0.088163722 
#>      gene133      gene134      gene135      gene136      gene137      gene138 
#> -5.308028594  0.079302484 -1.574800677  0.838683887 -2.020428749  0.398047144 
#>      gene139      gene140      gene141      gene142      gene143      gene144 
#>  0.309960972 -1.105963162  0.085025989 -0.015508483 -0.910475540 -0.177777676 
#>      gene145      gene146      gene147      gene148      gene149      gene150 
#>  0.879123434 -0.626760288 -0.422163814 -0.352155624 -0.071464837 -0.151108903 
#>      gene151      gene152      gene153      gene154      gene155      gene156 
#> -0.211211748  0.757437072 -0.113909652 -0.044573202 -0.379224705  0.043463018 
#>      gene157      gene158      gene159      gene160      gene161      gene162 
#> -0.172677955 -1.208967865 -0.954411600  0.248054311 -0.191041364 -0.264339465 
#>      gene163      gene164      gene165      gene166      gene167      gene168 
#>  1.081825134 -0.564545435 -0.287617297  0.514900597 -0.516299390  0.343026461 
#>      gene169      gene170      gene171      gene172      gene173      gene174 
#>  0.359539033  0.002826349  1.966513978 -0.005099595 -0.022020332 -0.418350459 
#>      gene175      gene176      gene177      gene178      gene179      gene180 
#>  1.033760806 -0.114956170 -0.290113425 -0.264298061  0.088407594  1.708597897 
#>      gene181      gene182      gene183      gene184      gene185      gene186 
#>  0.346155981  1.128017842  0.692499670  0.032467791 -0.675453825  1.282242660 
#>      gene187      gene188      gene189      gene190      gene191      gene192 
#> -0.050294576 -0.016456977 -0.269830953  2.604537853 -1.572190346 -0.632763011 
#>      gene193      gene194      gene195      gene196      gene197      gene198 
#>  1.050187486 -0.060098061 -0.025864981 -1.975250501 -0.352373131 -0.564013543 
#>      gene199      gene200      gene201      gene202      gene203      gene204 
#> -0.070694531  2.002886234 -1.552197058 -0.750232421 -0.242458816  0.695399321 
#>      gene205      gene206      gene207      gene208      gene209      gene210 
#> -0.344427912  0.274773252  1.124446056  0.069091388 -0.064702137 -0.788906973 
#>      gene211      gene212      gene213      gene214      gene215      gene216 
#> -1.407602754 -0.136112762 -1.112413371 -0.893750115  0.350273885  0.128153895 
#>      gene217      gene218      gene219      gene220      gene221      gene222 
#>  1.520751877  0.065502227  0.998523026 -0.160957820  0.210162450 -0.235956247 
#>      gene223      gene224      gene225      gene226      gene227      gene228 
#> -0.059074175  0.516304933  0.875139605  1.005682399 -0.111251776 -0.191631470 
#>      gene229      gene230      gene231      gene232      gene233      gene234 
#>  0.189596836 -0.227074114  0.220864628 -0.367068354  0.142885428  0.163566575 
#>      gene235      gene236      gene237      gene238      gene239      gene240 
#> -0.309361501  0.093716242 -0.263739920 -0.434546827 -0.284481975 -0.913077268 
#>      gene241      gene242      gene243      gene244      gene245      gene246 
#> -0.393824947  1.745783368 -0.103139035  0.023671856  0.068432798  2.116937505 
#>      gene247      gene248      gene249      gene250      gene251      gene252 
#>  0.002205639  1.512189283 -0.038174105 -0.741269353 -0.603779903 -0.148523095 
#>      gene253      gene254      gene255      gene256      gene257      gene258 
#>  0.198317376 -0.613764019  0.626380554  0.057377496  0.604701361  0.629540044 
#>      gene259      gene260      gene261      gene262      gene263      gene264 
#>  0.035339655 -1.368036449  0.060994378  1.653525604 -0.354343838 -0.317950302 
#>      gene265      gene266      gene267      gene268      gene269      gene270 
#>  0.190663085  0.474069407 -0.495792095  0.087525954  0.532830427 -0.567247887 
#>      gene271      gene272      gene273      gene274      gene275      gene276 
#> -1.385963152  1.589517066  0.005217875 -0.066711804 -0.181874630  0.598528451 
#>      gene277      gene278      gene279      gene280      gene281      gene282 
#>  0.203817400 -0.155960889 -0.064805840 -0.244291396 -0.969195036 -0.021511092 
#>      gene283      gene284      gene285      gene286      gene287      gene288 
#> -0.124827906 -0.760454742  0.261239666 -1.567553726  0.407626880 -0.943835817 
#>      gene289      gene290      gene291      gene292      gene293      gene294 
#>  0.447794784  0.530608848  1.732841022  0.663545430 -2.916363485 -0.353380722 
#>      gene295      gene296      gene297      gene298      gene299      gene300 
#> -0.345734396 -0.197949862  0.954452837  0.327363803 -0.573764483 -0.049088193 
#>      gene301      gene302      gene303      gene304      gene305      gene306 
#>  0.103667162  0.772860061 -0.539717273  1.851152401 -0.285978507  1.360894047 
#>      gene307      gene308      gene309      gene310      gene311      gene312 
#>  1.102997809  0.027969944 -0.335768513 -0.599693789 -0.474708641  0.331346396 
#>      gene313      gene314      gene315      gene316      gene317      gene318 
#>  0.386733062  0.691539435 -0.168147376  0.291080242 -0.204509312 -0.457602029 
#>      gene319      gene320      gene321      gene322      gene323      gene324 
#>  0.556680175  1.560043727 -0.666739311  0.632170528 -0.255692626  0.351021182 
#>      gene325      gene326      gene327      gene328      gene329      gene330 
#>  1.182021258  0.819710466 -0.600048521 -1.185316674  0.722400761 -0.110984775 
#>      gene331      gene332      gene333      gene334      gene335      gene336 
#> -0.338742654  0.649056183 -0.339437138  0.167097785 -0.095236322  0.407179778 
#>      gene337      gene338      gene339      gene340      gene341      gene342 
#> -0.211501086 -0.442096903 -0.273290677  0.147645734  0.005595966  0.223472913 
#>      gene343      gene344      gene345      gene346      gene347      gene348 
#>  3.526418113  0.613447204  0.033552037  0.127677569 -0.025246384  0.066658952 
#>      gene349      gene350      gene351      gene352      gene353      gene354 
#> -0.339297395  0.439191215  0.001400679 -0.643728533 -0.459308017  0.570270707 
#>      gene355      gene356      gene357      gene358      gene359      gene360 
#>  0.243915510 -0.069043165  0.232521608 -0.355299649 -0.039408529 -0.360479538 
#>      gene361      gene362      gene363      gene364      gene365      gene366 
#> -1.736142309  0.164830120  2.079362599 -0.571597435 -1.256812764  0.037727791 
#>      gene367      gene368      gene369      gene370      gene371      gene372 
#> -0.091736780  0.401931525 -0.415677287  0.561256684  0.559642486  0.071692575 
#>      gene373      gene374      gene375      gene376      gene377      gene378 
#> -1.941313384 -0.323387987 -2.713687626 -0.826078088  0.249397049 -0.244746026 
#>      gene379      gene380      gene381      gene382      gene383      gene384 
#> -0.283451601  0.119365883 -0.097532228 -0.356695779 -0.253436503  0.819507715 
#>      gene385      gene386      gene387      gene388      gene389      gene390 
#> -0.910433120  0.021374804  1.527066941  0.033855889  0.015791169 -0.233504442 
#>      gene391      gene392      gene393      gene394      gene395      gene396 
#>  0.119614748 -0.164938700 -0.296979538 -0.093127757 -0.021132155  0.491994869 
#>      gene397      gene398      gene399      gene400      gene401      gene402 
#> -0.145829479  0.686998167  0.204491344  0.160315518  0.131712618 -0.187849250 
#>      gene403      gene404      gene405      gene406      gene407      gene408 
#> -1.612150315 -0.449001493 -0.251757546 -2.045377170  0.234922145 -0.008630738 
#>      gene409      gene410      gene411      gene412      gene413      gene414 
#>  0.500637203  0.596793530 -0.648646764 -0.643945566 -0.575365690 -0.703011034 
#>      gene415      gene416      gene417      gene418      gene419      gene420 
#> -0.084249971  0.135612844  0.741507111 -1.247380923  1.221105688  0.953846460 
#>      gene421      gene422      gene423      gene424      gene425      gene426 
#> -0.080808788  1.321816966 -0.372914066 -0.297366447 -0.121040082 -0.738872900 
#>      gene427      gene428      gene429      gene430      gene431      gene432 
#>  0.468016914  0.082134201 -0.804631064 -1.387111072  2.184589473  0.725717263 
#>      gene433      gene434      gene435      gene436      gene437      gene438 
#>  0.422307122 -0.190007817 -0.633274842  1.799386122 -0.247106746  0.771404414 
#>      gene439      gene440      gene441      gene442      gene443      gene444 
#> -0.127249984  2.752189132 -1.177057469 -1.513821848  1.325041014  0.496038474 
#>      gene445      gene446      gene447      gene448      gene449      gene450 
#>  0.082561683 -0.738817611 -0.411492177  0.085514620  0.054209271 -0.095728383 
#>      gene451      gene452      gene453      gene454      gene455      gene456 
#>  0.891404247  0.305273731 -0.176434621  0.337078305 -0.100461005 -2.897751839 
#>      gene457      gene458      gene459      gene460      gene461      gene462 
#>  0.313159880 -0.164508528  0.163120618  0.128532890 -0.422574982  0.324286153 
#>      gene463      gene464      gene465      gene466      gene467      gene468 
#> -0.067916354 -0.412055212 -0.416042385 -0.810261719 -1.018644707 -0.976821151 
#>      gene469      gene470      gene471      gene472      gene473      gene474 
#> -0.016709394  1.717376207 -0.836144387 -0.312432319 -0.139216197 -1.464665770 
#>      gene475      gene476      gene477      gene478      gene479      gene480 
#> -1.826636877  0.047033758  0.544302919 -0.358199492 -0.564986462  0.409127451 
#>      gene481      gene482      gene483      gene484      gene485      gene486 
#> -0.482583588  0.193536831  0.325698524 -0.367204722 -0.035104591  0.199600000 
#>      gene487      gene488      gene489      gene490      gene491      gene492 
#>  1.018929901  2.694371148 -0.524816660 -1.027440898  0.355939004 -0.022359185 
#>      gene493      gene494      gene495      gene496      gene497      gene498 
#> -0.323586382 -0.050761521  0.539950264 -0.377143708  0.290883371  0.057839095 
#>      gene499      gene500 
#> -2.761587382  0.734258598