Results tables

Results tables

resultsTables(object, ...)

# S4 method for DESeqResults
resultsTables(object, DESeqDataSet = NULL,
  alpha = NULL, lfcThreshold = NULL, return = c("tbl_df",
  "DataFrameList"), ...)

# S4 method for DESeqAnalysis
resultsTables(object, results,
  lfcShrink = TRUE, extra = TRUE, alpha = NULL,
  lfcThreshold = NULL, return = c("tbl_df", "DataFrameList"))

Arguments

object
Object.

DESeqDataSet
DESeqDataSet or NULL.

alpha
numeric(1) or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

lfcThreshold
numeric(1) or NULL. Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

return
character(1). Type of data frame to return as a list. Uses match.arg().

DataFrameList: Returns DataFrameList with row names.

tbl_df: Returns list of tbl_df containing "rowname" column.

...
Additional arguments.

results
character(1) or integer(1). Name or position of DESeqResults.

lfcShrink
logical(1). Use shrunken log2 fold change (LFC) values.

extra
logical(1). Include row data and normalized counts from internal DESeqDataSet.

contrast
character(1) or NULL. Contrast name.

Value

list. Named list containing subsets of DESeqResults.

Details

Generate tables summarizing the differential expression, with subsets for differentially expressed genes (DEGs). DEG tables (i.e. everything except the all table), are arranged by adjusted P value.

Note

Updated 2019-08-20.

Do not apply post hoc log fold change cutoffs.

Tables

all: All genes, including genes without an adjusted P value. This table is unmodified, and the rows have not been re-arranged or subset. It is suitable for gene set enrichment analysis (GSEA).

up: Upregulated genes.

down: Downregulated genes.

both: Bidirectional DEGs (up- and down-regulated). This table can be used for overrepresentation testing but should NOT be used for GSEA.

Examples

data(deseq)

## DESeqAnalysis ====
x <- resultsTables(deseq, results = 1L)

#> condition_B_vs_A (shrunken LFC)

#> Returning with the sample names unmodified.

#> Joining row annotations.

#> Joining size factor adjusted normalized counts.

#> 156 differentially expressed genes detected.

#> 81 upregulated genes detected.

#> 75 downregulated genes detected.

names(x)

#> [1] "all"  "up"   "down" "both"

## DESeqResults ====
## Use of DESeqAnalysis is encouraged instead of this approach.
res <- results(deseq, results = 1L)

#> condition_B_vs_A (unshrunken LFC)

dds <- as(deseq, "DESeqDataSet")
x <- resultsTables(object = res, DESeqDataSet = dds)

#> Joining row annotations.

#> Joining size factor adjusted normalized counts.

#> 156 differentially expressed genes detected.

#> 81 upregulated genes detected.

#> 75 downregulated genes detected.

names(x)

#> [1] "all"  "up"   "down" "both"