DESeq2 Differential Expression Results Tables

DESeqResults object with DataFrame subsets and file path information.

Value

DESeqResultsTables, containing the original, unmodified DESeqResults DataFrame along with the corresponding differentially expressed genes and gene-level metadata (rowRanges).

Slots

results

DESeqResults. Original unmodified DESeqResults. Should contain all genes, including those with NA adjusted P values.

deg

list. Differentially expressed genes. Contains character vectors of genes that are upregulated (up) or downregulated (down). Values map to the rownames of the internal DESeqResults. These are genes that pass alpha and lfcThreshold cutoffs set in DESeq2::results() call.

counts

matrix. Normalized counts matrix.

rowRanges

GRanges. Row annotations.

sampleNames

character. Human-friendly sample names. Must contain names() that map to the colnames() of the DESeqDataSet.

metadata

list. Metadata. Contains file paths and information on whether we're writing locally or to Dropbox.

See also