Alpha level cutoff summary statistics

Quickly generate a summary table of various alpha level cutoffs.

alphaSummary(object, ...)

# S4 method for DESeqDataSet
alphaSummary(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), contrast = NULL, name = NULL)

Arguments

object
Object.

alpha
numeric. Multiple alpha cutoffs.

contrast
this argument specifies what comparison to extract from the object to build a results table. one of either:

a character vector with exactly three elements: the name of a factor in the design formula, the name of the numerator level for the fold change, and the name of the denominator level for the fold change (simplest case)

a list of 2 character vectors: the names of the fold changes for the numerator, and the names of the fold changes for the denominator. these names should be elements of resultsNames(object). if the list is length 1, a second element is added which is the empty character vector, character(). (more general case, can be to combine interaction terms and main effects)

a numeric contrast vector with one element for each element in resultsNames(object) (most general case)

If specified, the name argument is ignored.

name
the name of the individual effect (coefficient) for building a results table. Use this argument rather than contrast for continuous variables, individual effects or for individual interaction terms. The value provided to name must be an element of resultsNames(object).

...
Additional arguments.

Value

integer matrix.

Details

Use either contrast or name to specify the desired contrast.

Note

Updated 2019-08-20.

See also

Examples

data(deseq)

## DESeqDataSet ====
dds <- as(deseq, "DESeqDataSet")
resultsNames(dds)

#> [1] "Intercept"        "condition_B_vs_A"

alphaSummary(dds)

#> Intercept
#> Alpha levels: 0.1, 0.05, 0.01, 0.001, 1e-06

#>                0.1 0.05 0.01 0.001 1e-06
#> LFC > 0 (up)    81   64   47    30     6
#> LFC < 0 (down)  75   60   39    23     4
#> outliers [1]     2    2    2     2     2
#> low counts [2]   0    0   96   173     0

alphaSummary(dds, contrast = c("condition", "B", "A"))

#> condition B A
#> Alpha levels: 0.1, 0.05, 0.01, 0.001, 1e-06

#>                0.1 0.05 0.01 0.001 1e-06
#> LFC > 0 (up)    81   64   47    30     6
#> LFC < 0 (down)  75   60   39    23     4
#> outliers [1]     2    2    2     2     2
#> low counts [2]   0    0   96   173     0

alphaSummary(dds, name = "condition_B_vs_A")

#> condition_B_vs_A
#> Alpha levels: 0.1, 0.05, 0.01, 0.001, 1e-06

#>                0.1 0.05 0.01 0.001 1e-06
#> LFC > 0 (up)    81   64   47    30     6
#> LFC < 0 (down)  75   60   39    23     4
#> outliers [1]     2    2    2     2     2
#> low counts [2]   0    0   96   173     0