Differentially expressed genes

Differentially expressed genes

deg(object, ...)

# S4 method for DESeqResults
deg(
  object,
  alpha = NULL,
  lfcThreshold = NULL,
  direction = c("both", "up", "down")
)

# S4 method for DESeqAnalysis
deg(object, i, ...)

Arguments

object
Object.

alpha
numeric(1) or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

lfcThreshold
numeric(1) or NULL. Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

direction
character(1). Include "both", "up", or "down" directions.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

...
Additional arguments.

Value

character. Gene identifiers.

Note

Updated 2019-12-18.

Examples

data(deseq)

## DESeqAnalysis ====
x <- deg(deseq, i = 1L)

#> condition_B_vs_A (unshrunken LFC)

#> 38 differentially expressed genes detected (alpha: 0.01; lfc: 0).

head(x)

#> [1] "gene338" "gene214" "gene380" "gene028" "gene112" "gene401"