Results tables

Results tables

# S4 method for DESeqResults
resultsTables(object, return = c("tbl_df",
  "DataFrameList"))

# S4 method for DESeqAnalysis
resultsTables(object, results, rowData = TRUE,
  counts = TRUE, return = c("tbl_df", "DataFrameList"))

Arguments

object
Object.

return
character(1). Type of data frame to return in the list. Uses match.arg(). Note that DataFrame option will return with row names, whereas tbl_df option will return with "rowname" column.

results
character(1) or integer(1). Name or position of DESeqResults.

rowData
logical(1). Join the row annotations.

counts
logical(1). Join the size-factor adjusted normalized counts.

Value

list. Named list containing subsets of DESeqResults.

Details

Generate tables summarizing the differential expression, with subsets for differentially expressed genes (DEGs). DEG tables (i.e. everything except the all table), are arranged by adjusted P value.

Note

Do not apply post hoc log fold change cutoffs.

Tables

all: All genes, including genes without an adjusted P value. This table is unmodified, and the rows have not been re-arranged or subset. It is suitable for gene set enrichment analysis (GSEA).

up: Upregulated genes.

down: Downregulated genes.

both: Bi-directional DEGs (up- and down-regulated). This table can be used for overrepresentation testing but should NOT be used for GSEA.

Examples

data(deseq)

## DESeqAnalysis ====
x <- resultsTables(deseq, results = 1L)

#> DESeqResults: condition B vs A (shrunken LFC)

#> Returning with the sample names unmodified.

#> Joining row annotations.

#> Joining size factor adjusted normalized counts.

#> 171 differentially expressed genes detected.

#> 84 upregulated genes detected.

#> 87 downregulated genes detected.

print(x)

#> $all
#> # A tibble: 500 x 22
#>    rowname baseMean log2FoldChange lfcSE  pvalue    padj geneID geneName
#>    <chr>      <dbl>          <dbl> <dbl>   <dbl>   <dbl> <chr>  <fct>   
#>  1 gene001  148.            1.39   0.404 1.00e-4 1.27e-3 ENSG0… TSPAN6  
#>  2 gene002   34.0          -0.178  0.368 5.64e-1 7.24e-1 ENSG0… TNMD    
#>  3 gene003   10.6          -1.76   0.684 9.93e-4 7.71e-3 ENSG0… DPM1    
#>  4 gene004   81.2          -1.94   0.433 7.80e-7 2.77e-5 ENSG0… SCYL3   
#>  5 gene005   11.6           0.195  0.479 5.74e-1 7.34e-1 ENSG0… C1orf112
#>  6 gene006  252.           -0.0297 0.265 9.02e-1 9.40e-1 ENSG0… FGR     
#>  7 gene007   81.1           0.0880 0.293 7.41e-1 8.48e-1 ENSG0… CFH     
#>  8 gene008   64.1           0.529  0.432 1.25e-1 2.64e-1 ENSG0… FUCA2   
#>  9 gene009    5.12          0.186  0.555 5.83e-1 7.41e-1 ENSG0… GCLC    
#> 10 gene010    0.728        -0.0237 0.658 8.88e-1 9.37e-1 ENSG0… NFYA    
#> # … with 490 more rows, and 14 more variables: geneBiotype <fct>,
#> #   broadClass <fct>, sample01 <dbl>, sample02 <dbl>, sample03 <dbl>,
#> #   sample04 <dbl>, sample05 <dbl>, sample06 <dbl>, sample07 <dbl>,
#> #   sample08 <dbl>, sample09 <dbl>, sample10 <dbl>, sample11 <dbl>,
#> #   sample12 <dbl>
#> 
#> $up
#> # A tibble: 84 x 22
#>    rowname baseMean log2FoldChange lfcSE   pvalue    padj geneID geneName
#>    <chr>      <dbl>          <dbl> <dbl>    <dbl>   <dbl> <chr>  <fct>   
#>  1 gene095    123.            2.36 0.367 1.12e-11 1.86e-9 ENSG0… MPO     
#>  2 gene334     72.4           2.85 0.458 3.68e-11 4.57e-9 ENSG0… RWDD2A  
#>  3 gene323    195.            1.99 0.321 5.67e-11 5.64e-9 ENSG0… UBR7    
#>  4 gene494    113.            1.76 0.284 7.91e-11 6.55e-9 ENSG0… EIPR1   
#>  5 gene391    901.            1.62 0.276 7.30e-10 4.54e-8 ENSG0… PRDM11  
#>  6 gene145     93.3           2.16 0.410 1.36e- 8 7.53e-7 ENSG0… USH1C   
#>  7 gene078     87.0           1.72 0.334 3.50e- 8 1.74e-6 ENSG0… HOXA11  
#>  8 gene403     47.9           1.77 0.364 1.55e- 7 6.60e-6 ENSG0… RUNX3   
#>  9 gene248    153.            1.31 0.271 2.47e- 7 9.44e-6 ENSG0… METTL13 
#> 10 gene489    155.            1.58 0.367 2.44e- 6 7.59e-5 ENSG0… ARHGAP31
#> # … with 74 more rows, and 14 more variables: geneBiotype <fct>,
#> #   broadClass <fct>, sample01 <dbl>, sample02 <dbl>, sample03 <dbl>,
#> #   sample04 <dbl>, sample05 <dbl>, sample06 <dbl>, sample07 <dbl>,
#> #   sample08 <dbl>, sample09 <dbl>, sample10 <dbl>, sample11 <dbl>,
#> #   sample12 <dbl>
#> 
#> $down
#> # A tibble: 87 x 22
#>    rowname baseMean log2FoldChange lfcSE   pvalue     padj geneID geneName
#>    <chr>      <dbl>          <dbl> <dbl>    <dbl>    <dbl> <chr>  <fct>   
#>  1 gene427     44.0          -2.90 0.359 6.44e-17 3.20e-14 ENSG0… ERP44   
#>  2 gene419     27.8          -2.99 0.424 1.51e-13 3.75e-11 ENSG0… SLC45A4 
#>  3 gene352     25.7          -2.86 0.496 6.38e-10 4.53e- 8 ENSG0… SLC30A9 
#>  4 gene497     48.3          -2.17 0.455 1.59e- 7 6.60e- 6 ENSG0… ALG1    
#>  5 gene004     81.2          -1.94 0.433 7.80e- 7 2.77e- 5 ENSG0… SCYL3   
#>  6 gene285    116.           -1.35 0.307 1.85e- 6 6.13e- 5 ENSG0… ABHD5   
#>  7 gene023     67.0          -1.25 0.291 3.75e- 6 1.00e- 4 ENSG0… LAP3    
#>  8 gene148    238.           -1.12 0.258 3.83e- 6 1.00e- 4 ENSG0… TBXA2R  
#>  9 gene473     12.8          -2.56 0.648 5.90e- 6 1.33e- 4 ENSG0… SNX1    
#> 10 gene412    152.           -1.18 0.287 8.94e- 6 1.93e- 4 ENSG0… AQR     
#> # … with 77 more rows, and 14 more variables: geneBiotype <fct>,
#> #   broadClass <fct>, sample01 <dbl>, sample02 <dbl>, sample03 <dbl>,
#> #   sample04 <dbl>, sample05 <dbl>, sample06 <dbl>, sample07 <dbl>,
#> #   sample08 <dbl>, sample09 <dbl>, sample10 <dbl>, sample11 <dbl>,
#> #   sample12 <dbl>
#> 
#> $both
#> # A tibble: 171 x 22
#>    rowname baseMean log2FoldChange lfcSE   pvalue     padj geneID geneName
#>    <chr>      <dbl>          <dbl> <dbl>    <dbl>    <dbl> <chr>  <fct>   
#>  1 gene427     44.0          -2.90 0.359 6.44e-17 3.20e-14 ENSG0… ERP44   
#>  2 gene419     27.8          -2.99 0.424 1.51e-13 3.75e-11 ENSG0… SLC45A4 
#>  3 gene095    123.            2.36 0.367 1.12e-11 1.86e- 9 ENSG0… MPO     
#>  4 gene334     72.4           2.85 0.458 3.68e-11 4.57e- 9 ENSG0… RWDD2A  
#>  5 gene323    195.            1.99 0.321 5.67e-11 5.64e- 9 ENSG0… UBR7    
#>  6 gene494    113.            1.76 0.284 7.91e-11 6.55e- 9 ENSG0… EIPR1   
#>  7 gene352     25.7          -2.86 0.496 6.38e-10 4.53e- 8 ENSG0… SLC30A9 
#>  8 gene391    901.            1.62 0.276 7.30e-10 4.54e- 8 ENSG0… PRDM11  
#>  9 gene145     93.3           2.16 0.410 1.36e- 8 7.53e- 7 ENSG0… USH1C   
#> 10 gene078     87.0           1.72 0.334 3.50e- 8 1.74e- 6 ENSG0… HOXA11  
#> # … with 161 more rows, and 14 more variables: geneBiotype <fct>,
#> #   broadClass <fct>, sample01 <dbl>, sample02 <dbl>, sample03 <dbl>,
#> #   sample04 <dbl>, sample05 <dbl>, sample06 <dbl>, sample07 <dbl>,
#> #   sample08 <dbl>, sample09 <dbl>, sample10 <dbl>, sample11 <dbl>,
#> #   sample12 <dbl>
#>